requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
openai==1.30.0
gunicorn==21.2.0
requests==2.31.0
psycopg2-binary==2.9.9
